email-tone-classifier/pyproject.toml
[project]
name = "email-tone-classifier"
version = "0.1.0"
description = "Learn Langfuse + Promptfoo with Anthropic Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "langfuse>=2.50.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pylama>=8.4.0",
    "isort>=5.13.0",
    "black>=24.0.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pylama]
linters = "pycodestyle,pyflakes"
max_line_length = 88

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
